@@ -1,20 +1,18 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import './hero.css';
 
-class Hero extends Component {
-  render() {
-    return (
-      <section className="hero is-info">
-        <div className="hero-body">
-          <div className="container">
-            <h1 className="title">{this.props.children}</h1>
-          </div>
+const Hero = ({ children }) => {
+  return (
+    <section className="hero is-info">
+      <div className="hero-body">
+        <div className="container">
+          <h1 className="title">{children}</h1>
         </div>
-      </section>
-    );
-  }
-}
+      </div>
+    </section>
+  );
+};
 
 Hero.propTypes = {
   children: PropTypes.node.isRequired
