@@ -1,31 +1,29 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import './modal.css';
 
-class Modal extends Component {
-  render() {
-    return (
-      <div className={`modal ${this.props.shown ? 'is-active' : ''}`}>
-        <div className="modal-background" onClick={this.props.hideModal} />
-        <div className="modal-content">
-          <div className="box">
-            {printHeaderIfNeeded(this.props)}
-            <section className="content">{this.props.body}</section>
-          </div>
+const Modal = ({ shown, body, hideModal, title }) => {
+  return (
+    <div className={`modal ${shown ? 'is-active' : ''}`}>
+      <div className="modal-background" onClick={hideModal} />
+      <div className="modal-content">
+        <div className="box">
+          {printHeaderIfNeeded({ title })}
+          <section className="content">{body}</section>
         </div>
       </div>
-    );
-  }
-}
+    </div>
+  );
+};
 
-function printHeaderIfNeeded(props) {
-  if (!props.title) {
+function printHeaderIfNeeded({ title }) {
+  if (!title) {
     return;
   }
 
   return (
     <header>
-      <p className="title">{props.title}</p>
+      <p className="title">{title}</p>
     </header>
   );
 }
