@@ -1,20 +1,15 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import './resultBox.css';
 
-class ResultBox extends Component {
-  render() {
-    return (
-      <div
-        className={`result-box ${this.props.className}`}
-        style={this.props.style}
-      >
-        {this.props.children}
-      </div>
-    );
-  }
-}
+const ResultBox = ({ className, children, style }) => {
+  return (
+    <div className={`result-box ${className}`} style={style}>
+      {children}
+    </div>
+  );
+};
 
 ResultBox.propTypes = {
   children: PropTypes.node,
