@@ -1,16 +1,14 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import './section.css';
 
-class Section extends Component {
-  render() {
-    return (
-      <div className="section">
-        <div className="container">{this.props.children}</div>
-      </div>
-    );
-  }
-}
+const Section = ({ children }) => {
+  return (
+    <div className="section">
+      <div className="container">{children}</div>
+    </div>
+  );
+};
 
 Section.propTypes = {
   children: PropTypes.node.isRequired
